Refactor all modules to use float instead of double
Changed all relevant function signatures, implementations, and template instantiations from double to float for consistency and improved hardware compatibility. Updated normalization, softmax, elementwise multiplication, and transformer logic. Also refactored Phi model code to remove unnecessary batch dimensions, clarified rotary embedding, and added missing rotary parameter computation.

+++ modules/elementwise_mul.cpp
@@ -1,7 +1,7 @@
 #include "transformer.hpp"
 
 template <int M, int N, class T>
-void elementwise_mul(T out[M][N], T in[M][N], double in_scaling) {
+void elementwise_mul(T out[M][N], T in[M][N], float in_scaling) {
     for (int i = 0; i < M; i++) {
         for (int j = 0; j < N; j++) {
             out[i][j] = in[i][j] * in_scaling;
