+++ modules/softmax.cpp
@@ -2,15 +2,15 @@
 #include <cmath>
 
 template <int M, int N>
-void softmax(double out[M][N], double in[M][N]) {
+void softmax(float out[M][N], float in[M][N]) {
     #pragma HLS INTERFACE port=return mode=s_axilite
     #pragma HLS INTERFACE port=out mode=bram
     #pragma HLS INTERFACE port=in mode=bram
 
     for (int i = 0; i < M; i++) {
-        double max_val = in[i][0];
+        float max_val = in[i][0];
         maxval_1D(&max_val, in[i]);
-        double sum_exp = 0;
+        float sum_exp = 0;
         for (int j = 0; j < N; j++) {
             out[i][j] = exp(in[i][j] - max_val);
             sum_exp += out[i][j];
